refactor: api함수정리
