--- src/widgets/detailed-schedule/index.ts
@@ -1,1 +1,0 @@
-export { default as ListLayout } from './list-layout'
