--- src/widgets/detailed-schedule/list-layout.tsx
@@ -1,3 +1,0 @@
-export default function ListLayout() {
-  return
-}
